Created a BotTesting project to see if manal bots would work. They don't.
Referencing Helpers (our own dll) creates a problem here too, so it seems the raw files are necessary. However, it also turns out that even when they're in the same project and Robocode lets us load the robot and bring it into battle, it still won't start, and robocode seems unable to get contact with it. Requires further investigation.